rjv3: improve reliability when parsing properties
According to issues, there are more than expected cases where a property
has inaccurate length fields, which could confuse the parser and lead to
failure finding the echo key.

For example, #18 pointed out that a property with header2.type == 0
could have a undersized length field, and #58 found one with an invalid
length which is shorter than the header itself (and caused potential
out-of-bound access in the parser).

This commit addresses all known cases by far.